Factor out valid json response data provider
