+++ tests/Functional/Client/CreateFrontendTokenTest.php
@@ -4,16 +4,17 @@ declare(strict_types=1);
 
 namespace SmartAssert\UsersClient\Tests\Functional\Client;
 
-use GuzzleHttp\Psr7\Response;
 use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Message\ResponseInterface;
 use SmartAssert\UsersClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
 use SmartAssert\UsersClient\Tests\Functional\DataProvider\NetworkErrorExceptionDataProviderTrait;
+use SmartAssert\UsersClient\Tests\Functional\DataProvider\ValidJsonResponseDataProviderTrait;
 
 class CreateFrontendTokenTest extends AbstractClientTest
 {
     use NetworkErrorExceptionDataProviderTrait;
     use InvalidJsonResponseExceptionDataProviderTrait;
+    use ValidJsonResponseDataProviderTrait;
 
     /**
      * @dataProvider networkErrorExceptionDataProvider
@@ -44,30 +45,5 @@ class CreateFrontendTokenTest extends AbstractClientTest
         $actual = $this->client->createFrontendToken('email', 'password');
 
         self::assertEquals($expected, $actual);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function validJsonResponseDataProvider(): array
-    {
-        return [
-            'created' => [
-                'httpFixture' => new Response(
-                    200,
-                    ['content-type' => 'application/json'],
-                    (string) json_encode([
-                        'key1' => 'value1',
-                        'key2' => 'value2',
-                        'key3' => 'value3',
-                    ])
-                ),
-                'expected' => [
-                    'key1' => 'value1',
-                    'key2' => 'value2',
-                    'key3' => 'value3',
-                ],
-            ],
-        ];
     }
 }
